Refactor adapt output plugin functions into separate files
To help out any poor coders who have to read them...
